Done!
I've been working on this for the past 72 hours straight and my brain is now officially mush. I didn't get everything in the assignment done, mostly because I couldn't figure it out, but you know what? I'm really happy with how much I DID get done! I don't know if it matters, but I wrote down all of the requirements I could find in the assignment and checked off what I got done, if it helps with grading. I'm gonna go sleep now. Regardless of the grade, please know I had a lot of fun in the last half of making this!